webdriver: Wait for navigation to complete when creating session with all `WebView`s closed (#39040)
In #37410 for webdriver mode, we introduced mechanism to keep program
alive when all webview are closed to comply with spec. But the test
fails with 90% probability, as the `browsing_context` is only registered
in constellation's `browsing_contexts` map when documents are ready.
This PR waits for navigation to complete without verifying browsing
context existence (as it's guaranteed from context) in this special case
by refactoring existing code.

Testing:
https://github.com/servo/servo/blob/4571cc1b3b17b3a759260bdfdb2e163576b313e7/tests/wpt/tests/webdriver/tests/classic/close_window/close.py#L87
is no longer flaky. Previously, it has 90% probability to fail. Now it
always pass.

Signed-off-by: Euclid Ye <euclid.ye@huawei.com>